Ban unwraps
Add a blanket ban on unwrap() in the linter script. If we have something
we think can never be an error, we should be saying why. This allows us
to use unwrap() for quick prototype code with no risk of accidentally
leaving it in on final commit

Fix any uses of unwraps that remained in the codebase