Refactor to re-use readdir to resolve paths
Modifying filesystem layout is currently quite a pain. First we have to
add the item to the directory listing, then when we interact with a
path, we have to try to guess the code that generated it with some
matching heuristics.

This is stupid.

1. We are throwing away useful data (e.g. item id -> item name mapping)
2. It's a lot of extra work to write the matching functions

Replace readdir()'s iterators with iterators of PathPurpose + name.
These can be easily mapped to the previous DirEntries (by just using the
name), but can now be used to parse the path as well

Some side effects of the change
* We now need a FileType::File. We now end up calling our directory
  listing in order to parse path, which means we can no longer crash
  when we are handed a passthrough path
* parse_path errors are now a more complex type, which means we cannot
  just rely on the std::io::Error code in the c bindings path.