Add custom filters
Currently the only way to view items is by looking at /items/id. This is
not an effective way to browse anything. What we really want is to group
items by some criteria, and browse them by name that way

In the case of using this to track todos, I might want to show a folder
that presents "projects". However, ideally the code base doesn't know
about the concept of a project, as I am hoping that we can use this as a
generalized document storage system.

If we create a relationship "parent" <-> "child". We could create
"projects" by assigning each todo for a project as a child of a parent.
If we then want to visualize projects, we could package them as an SQL
query that returns items that have no parents.

Now if we want to avoid the concept of a project from leaking into the
codebase, we need a way for a user to specify what item filters they
want to see, and how to name them.

This has a nice side effect of being more flexible in the future.

* Add the concept of filters.
* Show filters in the root of the mounted filesystem
* Filters are stored in the database as collections of rules. Each rule
  gets mapped to another WHERE statement in the sql query
  * At this point we only support a single filter, the one we need for
    showing items with no parents
  * Filters are stored as a table of id -> name, and a set of rules that
    map back to that ID. At the time of writing, we imagine that each
    enum variant will result in a new db table.