Add binary to create items to fs
Before this commit, there's not a way to create items from within the
mounted filesystem. This is an important distinction from the db tool
which lets us modify the database directly, and show those changes on
the filesystem

The filesystem API should look like interacting with paths on the
filesystem, not by IDs of things in the database. We could take
inspiration from sysfs and try something like echo "name" > new, but
then how do we retrieve which item was created?

Instead, I think it makes sense to bundle executables to do tasks in the
filesystem itself. I have not seen this pattern before, but it seems
like the simplest thing to get what I want

This binary needs some way to interact with the database, but the
mounted filesystem does not really know where the source database is,
and I don't think it really should. The filesystem is the way we
interact with the data, and so we should add options there that provide
what we need.

* Add a special file to the filesystem that has an API on it. We can
  issue CreateItem requests by writing a json blob to the file, and see
  the result by reading it back
* Writes/reads are tracked at a per connection level. E.g. if I open the
  file twice at once, write and read, I won't get data associated with
  the other file handle
* Add a new binary "create-item" that just issues a CreateItem request
  and prints the resulting path. The vision is that we will add more
  binaries for more tasks, and external scripts can stitch those
  together to create items with project specific relationships etc.