Add relationship id as metadata file
This ends up simplifying some scripting around creating/using
relationships in a future commit. ./create-... binaries return paths to
items, not their IDs. This is useful if you want to just jump to the
thing you made, but is not great if you want to work with that ID to
bind relationships etc.

We could parse out the relationship id from the path, but it's nicer to
be able to `$(cat returned_path/id)` to get it