Add db schema support for item specific filters
We want to conditionally show filters as a subfolder of projects. This
enables a use case of having something like

/projects
  /some_project
    /open_tickets

open_tickets can be some form of complex filter. E.g. "items that are
children of some_project, but not children of "finished".

For this to work we need to generalize the previous "filter" concept and
turn it into a "condition set". Condition sets can be run in either the
context of a "root filter" or an "item filter"

"root filter" refers to the previous filter concept. Filters that show
up in the root directory of the file system

"item filters" are more complex. These have 1 set of conditions to see
if the item should show the filter as a subfolder, and another set of
conditions to actually generate the list of items

Add a root_filters and item_filters table to capture these ideas. The
root_filters table contains a single set of condition set ids that say
which condition sets should be activated for the root folder, the
item_filters is similar except it stores 2 condition set ids

Note that we do not implement any actual usage of the item_filters yet,
this is just a schema upgrade + whatever changes need to made to keep
the previous usage working

DB upgrade implemented as a sequence of functions that change DB state
and update the user_version pragma. We check the pragma and skip
unnecessary upgrade fns